Stream 更新
